requirements.txt
langchain==0.0.316
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.34.0
openai==0.28.0
faiss-cpu==1.8.0
altair==5.3.0
tiktoken==0.7.0
youtube_transcript_api==0.6.2
tiktoken==0.8.0
pillow==10.3.0
